refactor: 로또 번호 선택 전략 클래스가 많아져 별도의 패키지로 분리
