+++ src/main/java/lotto/domain/lotto/pick/CombinedNumberPickStrategy.java
@@ -1,10 +1,11 @@
-package lotto.domain.lotto;
+package lotto.domain.lotto.pick;
 
 import static lotto.utils.Validations.requireNotNull;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Stream;
+import lotto.domain.lotto.LottoNumbers;
 
 public class CombinedNumberPickStrategy implements NumberPickStrategy {
     final List<NumberPickStrategy> strategies;
