+++ src/main/java/lotto/domain/lotto/pick/NumberPickStrategy.java
@@ -1,6 +1,7 @@
-package lotto.domain.lotto;
+package lotto.domain.lotto.pick;
 
 import java.util.stream.Stream;
+import lotto.domain.lotto.LottoNumbers;
 
 @FunctionalInterface
 public interface NumberPickStrategy {
