+++ src/test/java/lotto/domain/lotto/pick/CombinedNumberPickStrategyTest.java
@@ -1,9 +1,10 @@
-package lotto.domain.lotto;
+package lotto.domain.lotto.pick;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
 import java.util.stream.Stream;
+import lotto.domain.lotto.LottoNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
